Quit timer (Debian patch, updated)
This is the Debian quit timer patch, but with -N and -s options de-coupled.

-N only does a shutdown of net out, as it used to.
In the case of a tcp -N client and server w/out -k,
client ^D causes the server to shut down due to an asymmetry in the code.
Server shutdown then causes the client to shut down.
This is the only case where -N appears to operate like -q0.

Implementation detail: there is a new ctrld_seen flag to explicitly record ^D
(or end of file).