Delete temporary AF_UNIX socket on exit
There was code to delete the AF_UNIX socket of a client whether temporary or not
but this relied on readwrite() returning. Some recent enhancements can make
readwrite() exit directly e.g. -q0 then ^D, so make sure to at least delete
a socket in /tmp. With the signal handler introduced in 3569cd0e, the socket is
also deleted on control-C.

The quit() function had the wrong signature for a signal handler: fix.